+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - correctly loading input history for all the opened projects not just active one
 - renamed worktree rebase button label to better describe the action
 - improved conflict detection during worktree merge and rebase operations
+- preventing auto-scroll pause on touch scroll up
 
 ## [0.77.0]